requirements.txt
celery
django
djangorestframework
Pillow
psycopg2-binary
python-dotenv
redis
requests
torch
torchvision